// src/api/model/testProfile.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { TestProfileTopicsItem } from './testProfileTopicsItem';

export interface TestProfile {
  gradePrediction?: number;
  percentageCourse?: number;
  percentageTests?: number;
  topics?: TestProfileTopicsItem[];
  userProfileImage?: string;
}